Added Machine Learning course project files
Understanding consumer behaviour helps businesses reduce waste, increase
operational efficiency, and concentrate resources on high-value areas. The performance of two
popular classification algorithms, Random Forest and Logistic Regression, in solving the binary
classification problem of determining if a particular customer's demographic and behavioural
characteristics suggest a likelihood of purchase is assessed in this study. 

The two tasks are customer churn prediction, which is important for the business to retain the high-
value clients, and financial forecasting models, where the high reliability of models is very much in
demand while making decisions in the banking and investment sectors.