Add files via upload
This study assesses the efficacy of two well-known clustering methods,
K-Means and DBSCAN are used on the provided dataset. Finding the best clustering algorithm for customer segmentation based on metrics like the
Silhouette Score and spotting significant trends that might inform data-driven business decisions
are the goals of this research.